# src/modules/apiDoc/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Social media",
        "description": "One simple social media project that developed by Ehsan to getting to know about developing Node js apps.",
        "version": "1.0.0"
    },
    "components":{
      "securitySchemes":{
        "Authorization":{
        "type":"apiKey",
        "name":"Authorization",
        "sheme":"bearer",
        "in":"header",
        "description":"User token authorization"
        }
      }
    },
...
    "paths": {
...
        "/pages/{pageID}/": {
            "get": {
                "tags": [
                    "Page"
                ],
                "summary": "Rendering a user page",
                "parameters": [
                    {
                        "name": "pageID",
                        "in": "path",
                        "required": "true",
                        "type": "string"
                    }
                ]
            }
        },
...
    }
}